pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Hacktober-Fest-2025-Pixly"
version = "0.1.0"
description = "Pixly is an AI game assistant overlay that helps you answers queries about a game without needing to leave it."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "customtkinter>=5.2.2",
    "fastapi>=0.117.1",
    "keyboard>=0.13.5",
    "requests>=2.32.5",
    "uvicorn>=0.37.0",
    "google-generativeai>=0.8.3",
    "pillow>=10.0.0",
    "dotenv>=0.9.9",
    "psutil>=5.9.0",
    "pywin32>=306",
    "cryptography>=41.0.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.2",
    "beautifulsoup4>=4.12.0",
    "pandas>=2.0.0",
    "lxml>=4.9.0",
    "pytest>=8.4.2",
    "agno>=2.1.8",
    "google-genai>=1.45.0",
    "sqlalchemy>=2.0.44",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.24.0",
    "pytest-xdist>=3.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
    "--cov=backend",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
    "api: API endpoint tests",
]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]
exclude = ["tests*", "vector_db*", "games_info*"]
